+++ tests/input/cannot-enumerate-pytest-fixtures/import_corrupted_module.py
@@ -1,0 +1,5 @@
+from no_such_package import fixture
+
+
+def test_something(fixture):
+    pass