+++ tests/input/cannot-enumerate-pytest-fixtures/no_such_package.py
@@ -1,0 +1,10 @@
+import pytest
+import this_is_invalid  # makes pytest fail
+
+
+@pytest.fixture
+def fixture():
+    pass
+
+def test_something(fixture):
+    pass